TOC Generation, Screen size Indicator Utility (#8)
* remove remarktoc; unset export

* add some markdown deps

* Add header component

* Add screen size indicator utility component

* add middleware to capture pathname and set xurl header

* TOC

The pathname will be retrieved by the `headers` utility from `next/headers`. It will be used to determine which post the user has requested. The toc.ts lib will then load the file and generate a TOC tree. This tree is then populated into the gui.

* Add webpack copyplugin

the md parsing logic does not work in prod because prod is built differently,
This commit adds a build step to copy the src directory to the build directory

* purge env file

* not found test

* not found test

* test show deploy dir

* probe .next dir

* probe vercel

* add vercel.json

* remove vercel.json